pyproject.toml
[tool.poetry]
name = "pytest_demo"
version = "0.1.0"
description = ""
authors = ["pythoninthegrass <4097471+pythoninthegrass@users.noreply.github.com>"]
license = "Unlicense"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
blinker = ">=1.4"
click = ">=7.0"
flask = ">=1.0.2"
flask-flatpages = ">=0.8.1"
flask-mail = ">=0.9.1"
flask-sqlalchemy = "^3.1.1"
flask-wtf = ">=0.14.2"
frozen-flask = ">=0.15"
itsdangerous = ">=1.1.0"
jinja2 = ">=2.10.1"
markdown = ">=3.0.1"
markupsafe = ">=1.1.0"
playwright = "^1.43.0"
psutil = "^5.9.8"
pygments = ">=2.3.0"
python-decouple = "^3.8"
pyyaml = ">=4.2b1"
werkzeug = ">=0.15.3"
wtforms = ">=2.2.1"

[tool.poetry.group.dev.dependencies]
coverage = "^7.4.4"
hypothesis = {extras = ["cli"], version = "^6.100.1"}
icecream = "^2.1.3"
ipython = "^8.23.0"
poetry-plugin-export = "^1.7.1"
poetry-plugin-up = "^0.7.1"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.3"
pytest-cov = "^5.0.0"
pytest-datafiles = "^3.0.0"
pytest-xdist = "^3.5.0"
rich = "^13.7.1"
ruff = "^0.4.1"

[tool.ruff]
# Fix without reporting on leftover violations
fix-only = true

# Enumerate all fixed violations
show-fixes = true

# Indent width (default: 4)
indent-width = 4

# Black (default: 88)
line-length = 130

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    "dist",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    "__pycache__",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Assume Python 3.11.
target-version = "py311"

[tool.ruff.format]
# Use spaces instead of tabs
indent-style = "space"

# Use `\n` line endings for all files
line-ending = "lf"

# Set quote style for strings
quote-style = "preserve"

[tool.ruff.lint]
# Enable pycodestyle (`E`), Pyflakes (`F`), and isort (`I`) by default.
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I001",
    "I002"
]
# Skip unused variable rules
ignore = []
extend-ignore = ["D203", "E203", "E251", "E266", "E401", "E402", "E501", "F401", "F403"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]

# unfixable = []

[tool.ruff.lint.isort]
from-first = false
lines-after-imports = 1
lines-between-types = 0
no-sections = true

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
